Turn off all output processing
Because terminal translation on output side from （“\n”）to （“\r\n”), The terminal requires both of these characters in order to start a new line of text.
The carriage return moves the cursor back to the beginning of the current line, and the newline moves the cursor down a line, scrolling the screen if necessary.
We turn off it by bitflag 'OPOST','O' means it’s an output flag, and 'I' assume POST stands for “post-processing of output”.
But if turn off thisfature, the output of newline characters only moving the cursor
down, not to the left side of screen, so we also update the printf
replace "\n" by "\r\n".