Clear the screen
Needs to clean the screen when the program start, by use write() and STDOUT_FILENO come from <unistd.h>.

We write 4 byte to terminal the first bytes is '\x1b' is an escape
character or '27' in decimal the other three bytes are '[2J'.

The escape character always followed by '[' character, escape sequences
instruct terminal to do various text formatting tasks(coloring text,
moving the cursor around, and clearing parts of the screen).

Use 'J' command to clear screen, '2' is the command arguments, which
says to clear the entire screen, argument '1' says to clear the screen up to
where the cursor is, argument '0'(default) says clear the screen from the cursor up to end
of screen.
ref:https://vt100.net/docs/vt100-ug/chapter3.html#ED

We use VT100 escape sequences which is support very widely by modern
terminalemulators.
ref:https://en.wikipedia.org/wiki/VT100
ref:https://vt100.net/docs/vt100-ug/chapter3.html

If we wanted to support the maximum number of terminals out there, we could use the ncurses library, which uses the terminfo database to figure out the capabilities of a terminal and what escape sequences to use for that particular terminal.
ref:https://en.wikipedia.org/wiki/Terminfo
ref:https://en.wikipedia.org/wiki/Ncurses