Reposition the cursor
Because after <esc>[2J command clear the screen, the cursor will at
bottom of screen, so needs to reposition it at top-left corner.

We use escape 'H' command to position the cursor.
The 'H' command could takes two arguments: the row number and column
number at which to position the cursor.
ex: in 80x24 size terminal can use '<esc>[12;40H'

The default arguments for 'H' both happen to be 1(the Rows and cloumns
are numbered starting at 1 ,not 0).