Clear the screen on exit
We needs to clear the screen and reposition the cursor when program
exits, if program error out also needs to clear the screen.

Note: we could use "atexit()" cuntion to clear the screen, but this
function will also clean the error message printed by "die()" funciton
that we are hard to debug.