Tildes
Drawing the tildes(~) at beginning of any lines that come after the end
of the file being edited like vim does.

Add "editorDrawRows()" function to drawing the tildes means that row is
not part of the file and cna't contain any text, we are not sure how
many rows just draw 24 rows.

After finish the drawing, reposition the cursor.