Clear lines one at a time
Instead of clearing the entire screen before each refresh, it seems more
optimal to clear each line as we redraw them, so remote the '<esc>[2J'
use '<esc>[K' sequence at end of line we draw.

Note: 'K' command(https://vt100.net/docs/vt100-ug/chapter3.html#EL) erases part of the current line. Its argument is analogous to the J command’s argument: 2 erases the whole line, 1 erases the part of the line to the left of the cursor, and 0 erases the part of the line to the right of the cursor. 0 is the default argument, and that’s what we want, so we leave out the argument and just use <esc>[K.