The Home and End keys
When user press 'Home' and 'End' keys needs to move the sursor to left
and right of screen.
But there are many different escape sequences that could sent by these
keys, depending on OS, or terminal emulator, so need to handle all of
it.

Home: <esc>[1~, <esc>[7~, <esc>[H, or <esc>OH
End: <esc>[4~, <esc>[8~, <esc>[F, or <esc>OF

- Add 'Home' and 'End' key enum value to 'editorKey'
- Update "editorReadKey()" function to catch above escape key and return
  corresponding enum value.
- Update "editorProcessKeypress()" function to handle these two enum key
  and move cursor to left and right.